UDT and co giveup, reversal working, RSA implementation ok but need AES 256 to work nicely
    The RSA algorithm can only encrypt data that has a maximum byte length of the RSA key length in bits divided with eight minus eleven padding bytes, i.e. number of maximum bytes = key length in bits / 8 - 11.

So basicly you divide the key length with 8 -11(if you have padding). For example if you have a 2048bit key you can encrypt 2048/8 = 256 bytes (- 11 bytes if you have padding). So, either use a larger key or you encrypt the data with a symmetric key, and encrypt that key with rsa (which is the recommended approach).

That will require you to:

    generate a symmetric key
    Encrypt the data with the symmetric key
    Encrypt the symmetric key with rsa
    send the encrypted key and the data
    Decrypt the encrypted symmetric key with rsa
    decrypt the data with the symmetric key
    done :)

cf steps : https://stackoverflow.com/questions/10007147/getting-a-illegalblocksizeexception-data-must-not-be-longer-than-256-bytes-when#10007285